# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Theorema Aureum 143 — Certificate Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CertificateStatus = typeof CertificateStatus[keyof typeof CertificateStatus];


export const CertificateStatus = {
  CERTIFIED: 'CERTIFIED',
  AWAITING: 'AWAITING',
  LOCKED: 'LOCKED',
  DISCHARGED: 'DISCHARGED',
} as const;

export interface Certificate {
  /** e.g. M1, M2, ... M7 */
  moduleId: string;
  title: string;
  /** One-sentence summary of what this module certifies */
  claim: string;
  status: CertificateStatus;
  /** @nullable */
  sourceFile?: string | null;
  sourceSha: string;
  stdoutSha: string;
  /** SHA-256 digests of parent module outputs this module depends on */
  parentShas: string[];
  /** Position in the M1→M7 DAG (1-indexed) */
  dagPosition: number;
  /**
     * Object storage path of the uploaded PDF
     * @nullable
     */
  pdfObjectPath?: string | null;
  /**
     * Lean theorem name bound to this certificate
     * @nullable
     */
  leanBinding?: string | null;
  /** @nullable */
  notes?: string | null;
  updatedAt?: string;
}

export interface CertificateUpdate {
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface CertificateSummary {
  /** Master manifest SHA-256 (cat m1.out..m6.out | sha256sum) */
  masterSha: string;
  totalModules: number;
  certifiedCount: number;
  awaitingCount: number;
  dagSealed: boolean;
  pdfUploadedCount?: number;
  pdfTotal?: number;
}

export interface LeanVerification {
  /** Lean toolchain version string */
  toolchain: string;
  /** Date the verification log was produced */
  dateVerified: string;
  /** List of remaining axioms (empty when fully discharged) */
  axiomDebt: string[];
  /** Raw "does not depend on any axioms" lines from `lake env lean Verify.lean` */
  axiomLines: string[];
  /** Full raw contents of lean-proof/VERIFY.txt */
  content: string;
  /** ISO-8601 timestamp of the VERIFY.txt file's last modification (mtime) */
  lastModified?: string;
  /** Age in days between the file's mtime and the time the response was generated */
  ageDays?: number;
}

export interface LeanRebuildResult {
  /** True iff the rebuild script exited 0 and VERIFY.txt was refreshed */
  ok: boolean;
  /** Exit code of the rebuild script (or -1 if it could not be spawned) */
  exitCode: number;
  stdout: string;
  stderr: string;
  /** Wall-clock duration of the rebuild in milliseconds */
  durationMs: number;
  /**
     * High-level error message when the script could not be launched or completed (e.g. `lake` not installed, timeout)
     * @nullable
     */
  error?: string | null;
  /** Re-parsed verification log after a successful rebuild */
  verification?: LeanVerification | null;
}

export interface LeanRebuildCancelResult {
  /** True when a cancellation signal was successfully sent. */
  ok: boolean;
  message?: string;
  error?: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}