# lib/api-zod/src/generated/types/leanRebuildCancelResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Theorema Aureum 143 — Certificate Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface LeanRebuildCancelResult {
  /** True when a cancellation signal was successfully sent. */
  ok: boolean;
  message?: string;
  error?: string;
}